code/requirements.txt
autopep8==1.5.7
awscli==1.19.77
beautifulsoup4==4.9.3
bs4==0.0.1
cryptography==3.4.7
elasticsearch==7.12.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
joblib==1.0.1
jupyter==1.0.0
jupyter-core==4.6.3
matplotlib==3.4.2
mistune==0.8.4
mlxtend==0.18.0
news-please==1.5.21
newsapi-python==0.2.6
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.20.3
pandas==1.2.4
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pydantic==1.7.4
PyDispatcher==2.0.5
PyMySQL==1.0.2
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==19.0.2
queuelib==1.6.1
readability-lxml==0.8.1
regex==2021.4.4
requests==2.25.1
requests-file==1.5.1
rsa==4.7.2
s3transfer==0.4.2
scikit-learn==0.24.2
scipy==1.6.3
Scrapy==2.5.0
sklearn==0.0
spacy==3.0.6
spacy-legacy==3.0.6
SQLAlchemy==1.4.15
wordcloud==1.8.1

